sql: fix 2-param version of setval to not set isCalled
Release note (bug fix, backward-incompatible change): The 2-parameter
setval builtin previously caused the sequence to increment incorrectly
one extra time. This is fixed now. If you do need the sequence to
increment, then use `setval(seq, val, true)`.

+++ pkg/sql/sequence.go
@@ -249,7 +249,7 @@ func getLatestValueInSessionForSequenceHelper(
 
 // SetSequenceValueByID implements the tree.SequenceOperators interface.
 func (p *planner) SetSequenceValueByID(
-	ctx context.Context, seqID int64, newVal int64, isCalled bool,
+	ctx context.Context, seqID uint32, newVal int64, isCalled bool,
 ) error {
 	if p.EvalContext().TxnReadOnly {
 		return readOnlyError("setval()")
@@ -275,7 +275,7 @@ func (p *planner) SetSequenceValueByID(
 	p.sessionDataMutatorIterator.applyOnEachMutator(func(m sessionDataMutator) {
 		m.initSequenceCache()
 		if isCalled {
-			m.RecordLatestSequenceVal(uint32(seqID), newVal)
+			m.RecordLatestSequenceVal(seqID, newVal)
 		}
 	})
 	return nil
